feat(experiments): enhance ExperimentDetailPage and related components with improved insights handling and UI consistency
